code for assignment 2 - FoA, 2023 Sem2
result of the assignment: 20/20

The project was aimed to create a text generator that is trained on user input strings in stage 0. 

On stage 0: the program read statements from input, construct its frequency prefix automaton, and print basic information about the automaton to the output 
On stage 1: the program will replay the automaton to generate the most frequently entered suffix on stage 0 that follows the user input string (prefix). 
On stage 2: extend the program to compress the automaton while maintaining the functionality of the functionality of this program